Move HTTPClientStub to its own file
